Use small int Java types for small UInt Franca types
Updated Java and JNI implementations to use Short for Franca UInt8 and
Integer for Franca UInt16. Larger UInt types still use Long as before.

Resolves: APIGEN-1501

+++ examples/platforms/android/app/src/main/java/com/example/here/hello/app/BuiltInFragment.java
@@ -105,7 +105,7 @@ public final class BuiltInFragment extends Fragment {
         result.setText(String.valueOf(byteResult));
         break;
       case 1:
-        long resultUInt8 = HelloWorldBuiltinTypes.methodWithUint8(Long.valueOf(parameterText));
+        short resultUInt8 = HelloWorldBuiltinTypes.methodWithUint8(Short.valueOf(parameterText));
         result.setText(String.valueOf(resultUInt8));
         break;
       case 2:
@@ -113,7 +113,7 @@ public final class BuiltInFragment extends Fragment {
         result.setText(String.valueOf(shortResult));
         break;
       case 3:
-        long resultUInt16 = HelloWorldBuiltinTypes.methodWithUint16(Long.valueOf(parameterText));
+        int resultUInt16 = HelloWorldBuiltinTypes.methodWithUint16(Integer.valueOf(parameterText));
         result.setText(String.valueOf(resultUInt16));
         break;
       case 4:
